requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.34.0
line-bot-sdk>=3.0.0
google-api-python-client>=2.150.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.0.0
openpyxl>=3.1.0
pandas>=2.2.0
pyarrow>=16.0.0
tenacity>=8.2.0
apscheduler>=3.10.0
matplotlib>=3.9.0
orjson>=3.10.0
python-dotenv>=1.0.0
psutil>=6.0.0

# Vector DB & Embeddings
sentence-transformers>=2.2.0
chromadb>=0.3.21
torch>=2.0.0

# Local AI - Ollama (LLM inference)
ollama>=0.1.0

# Gemini API (Free tier: 1M context, 15 RPM, 1500 RPD)
google-generativeai>=0.3.0

# Multimodal support (FREE tools)
easyocr>=1.7.0
vosk>=0.3.21
soundfile>=0.12.0
networkx>=3.0

# Dynamic file extraction
pytesseract>=0.3.10
PyPDF2>=3.0.1
python-docx>=0.8.11
pillow>=9.5.0

# Grafana integration
grafana-client>=3.0.0

# Testing
pytest>=7.0.0

